Seção 9: DTOs e Data Annotations

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,15 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using ProEventos.Persistence;
-using ProEventos.Domain;
-using ProEventos.Persistence.Contextos;
 using ProEventos.Application.Contratos;
 using Microsoft.AspNetCore.Http;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -31,7 +26,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if(eventos == null) return NotFound("Nenhum evento encontrado");
+                if(eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -48,7 +43,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if(evento == null) return NotFound("Nenhum evento encontrado");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -65,7 +60,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if(evento == null) return NotFound("Eventos por tema não encontrados");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -77,12 +72,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         { 
             try
             {
                 var evento = await _eventoService.AddEventos(model);
-                if(evento == null) return BadRequest("Erro ao tentar adicionar eventos");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -94,12 +89,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         { 
             try
             {
                 var evento = await _eventoService.UpdateEventos(id, model);
-                if(evento == null) return BadRequest("Erro ao tentar adicionar eventos");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -115,9 +110,12 @@ namespace ProEventos.API.Controllers
         { 
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if(evento == null) return NoContent();
+
                  return await _eventoService.DeleteEventos(id) ?
                     Ok("Deletado") :
-                    BadRequest("Evento não deletado");
+                    throw new Exception("Algo de errado não está certo!");
                 
             }
             catch (Exception ex)
